refactor registration and stabilization functions
